+++ src/Fields/CheckboxField.php
@@ -50,6 +50,10 @@ class CheckboxField extends AbstractField implements SingleValueInterface, Input
             return $this->checkedByPost;
         }
 
+        if (null !== $this->getValueOverride()) {
+            return (bool) $this->getValueOverride();
+        }
+
         return $this->checked;
     }