+++ src/Fields/TextareaField.php
@@ -93,7 +93,7 @@ class TextareaField extends AbstractField implements SingleValueInterface, Place
                 $this->translate($attributes->getPlaceholder() ?: $this->getPlaceholder())
             )
             . '>'
-            . $this->getValue()
+            . htmlentities($this->getValue())
             . '</textarea>';
     }
 }